Refactored project to incorporate a build system
The project now uses Autotools as its build system, and running the
usual './configure', 'make', 'make install' results in the building of a
shared library that gets installed in /usr/local/lib by default, as well
as the genann.h header, which gets installed in /usr/local/include.

All of the usual configuration flexibility is present, including the
addition of gmp and mpfr simply as a demonstration of the
auto-configuration ability of the system. The build system looks for gmp
and mpfr if the user specifies --with-gmp or --with-mpfr and looks for
the installation. If found, it automatically links the target to those
libraries.

The usual variables are configurable (CC, CFLAGS, CPPFLAGS, LDFLAGS,
LIBS), and the configuration even allows specifying the activation
function through there with ACTIVATION_FUNCTION=SIGMOID, although LINEAR
and THRESHOLD are also options. Specifying these options defines the
necessary preprocessor symbol.

Running 'make check' builds the test, links it against the library, and
runs it.

Running 'make examples' builds all four examples.