Removed inline function declarations
The inline function definition was causing users who simply included the
header and source file into other projects to get linker errors because
the inline functions were in the source file as opposed to the header.

While my gut instinct was to move the inline functions to the header,
the definition of a not-insignificant number of macros in the source
file meant this would have been a non-trivial task. For the moment, I
think the path least-resistance that manages to fix the job is to simply
remove the inline function specifier.

I'll make a note to profile the builds with and without the inline
specifier once I refactor, so we get a better idea of what should
happen, but for now I think it's best to just fix the builds without
simply declaring the functions as 'extern' which defeats the whole
purpose of an inline specification in the first place.

This commit fixes issue #28.